feat(quick-add): stream partial LLM results
Switch Quick Add from generateObject to streamObject (Vercel AI SDK) so
parsed fields populate live as tokens arrive. The previous spinner now only
shows until the first token, then a skeleton-backed preview fills in
progressively (kind, then primary name, time, count). On stream finish we
build the final QuickAddParseData via the same buildDataFromLlm pipeline so
committed values match the prior non-streaming behaviour.

- Split lib/quick-add-generate.ts: extracted client-safe schema and
  post-processing into lib/quick-add-build.ts; quick-add-generate.ts now
  re-exports for the existing test imports.
- Added app/api/quick-add/stream/route.ts (auth + tenant + rate limit +
  streamObject + after() usage logging). Replaces the parseQuickAdd server
  action, which is removed.
- Updated components/quick-add-input.tsx to consume the stream via
  experimental_useObject; resolves dayId via ensureDayExists before the
  stream so buildDataFromLlm can run on finish without a second round trip.
- New components/quick-add-streaming-preview.tsx renders the partial
  schema with plain skeleton placeholders for fields not yet streamed.

Closes #172
