+++ components/quick-add-input.tsx
@@ -1,13 +1,14 @@
 'use client'
 
-import { useEffect, useState, useTransition, type FormEvent, useId } from 'react'
+import { useEffect, useRef, useState, useTransition, type FormEvent, useId } from 'react'
 import { useRouter, usePathname } from 'next/navigation'
 import { useTranslations } from 'next-intl'
 import { Loader2, Sparkles } from 'lucide-react'
 import Link from 'next/link'
 import { toast } from 'sonner'
-import { parseQuickAdd } from '@/app/actions/quick-add'
+import { experimental_useObject as useObject } from '@ai-sdk/react'
 import { ensureDayExists } from '@/app/actions/days'
+import { quickAddLlmSchema, buildDataFromLlm } from '@/lib/quick-add-build'
 import type { QuickAddParseData } from '@/lib/quick-add-types'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Drawer, DrawerContent, DrawerHeader, DrawerTitle } from '@/components/ui/drawer'
@@ -17,6 +18,7 @@ import { Label } from '@/components/ui/label'
 import { mutateWithOfflineQueue } from '@/lib/day-mutation-client'
 import { useTenant } from '@/lib/tenant-context'
 import { QuickAddReview, type QuickAddReviewPayload } from '@/components/quick-add-review'
+import { QuickAddStreamingPreview } from '@/components/quick-add-streaming-preview'
 import type { Activity, BreakfastConfiguration, Reservation } from '@/types/index'
 
 type Props = {
@@ -26,7 +28,10 @@ type Props = {
   disabled?: boolean
 }
 
-type View = { stage: 'input' } | { stage: 'review'; data: QuickAddParseData; raw: string }
+type View =
+  | { stage: 'input' }
+  | { stage: 'streaming' }
+  | { stage: 'review'; data: QuickAddParseData; raw: string }
 
 function useIsMobile() {
   const [isMobile, setIsMobile] = useState(false)
@@ -51,10 +56,42 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
   const [view, setView] = useState<View>({ stage: 'input' })
   const [error, setError] = useState<string | null>(null)
   const [saveError, setSaveError] = useState<string | null>(null)
-  const [isParsing, startParse] = useTransition()
   const [isSaving, startSave] = useTransition()
   const descId = useId()
   const errId = useId()
+
+  // dayId is resolved client-side via ensureDayExists before the stream is
+  // submitted, so it's available when buildDataFromLlm runs on finish.
+  const dayIdRef = useRef<string | null>(null)
+  // Capture the raw input at submit time so it survives the user editing the
+  // textarea while the stream is in flight.
+  const submittedTextRef = useRef<string>('')
+
+  const {
+    object: streamObj,
+    submit: submitStream,
+    isLoading,
+    error: streamError,
+    stop: stopStream,
+  } = useObject({
+    api: '/api/quick-add/stream',
+    schema: quickAddLlmSchema,
+    onFinish({ object }: { object: unknown }) {
+      const parsed = quickAddLlmSchema.safeParse(object)
+      const dayId = dayIdRef.current
+      if (!parsed.success || !dayId) {
+        setView({ stage: 'input' })
+        if (!error) setError(t('parseFailed'))
+        return
+      }
+      const data = buildDataFromLlm(parsed.data, dayId, contextDate)
+      setView({ stage: 'review', data, raw: submittedTextRef.current })
+    },
+    onError(err: Error) {
+      setError(err?.message || t('parseFailed'))
+      setView({ stage: 'input' })
+    },
+  })
 
   useEffect(() => {
     if (!open) {
@@ -62,26 +99,36 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
       setView({ stage: 'input' })
       setError(null)
       setSaveError(null)
+      dayIdRef.current = null
+      submittedTextRef.current = ''
+      if (isLoading) stopStream()
     }
+    // We intentionally do not depend on isLoading/stopStream — only run on close.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [open])
 
   function close() {
     onOpenChange(false)
   }
 
-  function handleParse(e: FormEvent) {
+  async function handleParse(e: FormEvent) {
     e.preventDefault()
     const v = text.trim()
-    if (!v || isParsing) return
+    if (!v || isLoading) return
     setError(null)
-    startParse(async () => {
-      const r = await parseQuickAdd(v, contextDate)
-      if (!r.success) {
-        setError(r.error)
-        return
-      }
-      setView({ stage: 'review', data: r.data, raw: v })
-    })
+    dayIdRef.current = null
+    submittedTextRef.current = v
+    setView({ stage: 'streaming' })
+
+    // Resolve dayId up-front so buildDataFromLlm has it on stream finish.
+    const ensured = await ensureDayExists(contextDate)
+    if (!ensured.success) {
+      setError(ensured.error)
+      setView({ stage: 'input' })
+      return
+    }
+    dayIdRef.current = ensured.data.id
+    submitStream({ input: v, contextDate })
   }
 
   function handleBack() {
@@ -175,10 +222,26 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
     })
   }
 
+  // Surface stream errors to the inline error region.
+  useEffect(() => {
+    if (streamError && !error) {
+      setError(streamError.message || t('parseFailed'))
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [streamError])
+
   const isAiNotConfigured = Boolean(error?.includes('AI is not configured'))
+  const isParsing = view.stage === 'streaming'
+  // Spinner only until the first token arrives.
+  const showSpinner = isParsing && !streamObj
 
   const inputBody = (
-    <form onSubmit={handleParse} className="space-y-3">
+    <form
+      onSubmit={(e) => {
+        void handleParse(e)
+      }}
+      className="space-y-3"
+    >
       <p id={descId} className="text-muted-foreground text-sm">
         {t('description', { contextDate })}
       </p>
@@ -229,6 +292,36 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
     </form>
   )
 
+  const streamingBody =
+    view.stage === 'streaming' ? (
+      <div className="space-y-3">
+        {showSpinner ? (
+          <div
+            className="text-muted-foreground flex items-center justify-center gap-2 py-12 text-sm"
+            role="status"
+            aria-live="polite"
+          >
+            <Loader2 className="h-4 w-4 animate-spin" />
+            {t('parsing')}
+          </div>
+        ) : (
+          <QuickAddStreamingPreview partial={streamObj} label={t('parsing')} />
+        )}
+        <div className="flex justify-end gap-2 pt-1">
+          <Button
+            type="button"
+            variant="outline"
+            onClick={() => {
+              stopStream()
+              setView({ stage: 'input' })
+            }}
+          >
+            {t('cancel')}
+          </Button>
+        </div>
+      </div>
+    ) : null
+
   const reviewBody =
     view.stage === 'review' ? (
       <QuickAddReview
@@ -241,7 +334,10 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
       />
     ) : null
 
-  const title = view.stage === 'input' ? t('title') : t('reviewTitle')
+  const title = view.stage === 'input' || view.stage === 'streaming' ? t('title') : t('reviewTitle')
+
+  const body =
+    view.stage === 'input' ? inputBody : view.stage === 'streaming' ? streamingBody : reviewBody
 
   if (isMobile) {
     return (
@@ -259,9 +355,7 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
               {title}
             </DrawerTitle>
           </DrawerHeader>
-          <div className="max-h-[75vh] overflow-y-auto px-4 pb-6">
-            {view.stage === 'input' ? inputBody : reviewBody}
-          </div>
+          <div className="max-h-[75vh] overflow-y-auto px-4 pb-6">{body}</div>
         </DrawerContent>
       </Drawer>
     )
@@ -285,7 +379,7 @@ export function QuickAddInput({ open, onOpenChange, contextDate, disabled }: Pro
             {title}
           </DialogTitle>
         </DialogHeader>
-        {view.stage === 'input' ? inputBody : reviewBody}
+        {body}
       </DialogContent>
     </Dialog>
   )
